Move FilterMetadata into its own module
