+++ tiledb/arrow/src/attribute.rs
@@ -3,41 +3,11 @@ use std::collections::HashMap;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use tiledb::context::Context as TileDBContext;
-use tiledb::filter::{FilterData, FilterListBuilder};
+use tiledb::filter::FilterListBuilder;
 use tiledb::{error::Error as TileDBError, fn_typed, Result as TileDBResult};
 
 use crate::datatype::{arrow_type_physical, tiledb_type_physical};
-
-/// Encapsulates TileDB filter data for storage in Arrow Field metadata
-#[derive(Deserialize, Serialize)]
-pub struct FilterMetadata {
-    filters: Vec<FilterData>,
-}
-
-impl FilterMetadata {
-    pub fn new(
-        filters: &tiledb::filter_list::FilterList,
-    ) -> TileDBResult<Self> {
-        Ok(FilterMetadata {
-            filters: filters
-                .to_vec()?
-                .into_iter()
-                .map(|f| f.filter_data())
-                .collect::<TileDBResult<Vec<FilterData>>>()?,
-        })
-    }
-
-    /// Updates a FilterListBuilder with the contents of this object
-    pub fn apply<'ctx>(
-        &self,
-        mut filters: FilterListBuilder<'ctx>,
-    ) -> TileDBResult<FilterListBuilder<'ctx>> {
-        for filter in self.filters.iter() {
-            filters = filters.add_filter_data(filter.clone())?;
-        }
-        Ok(filters)
-    }
-}
+use crate::filter::FilterMetadata;
 
 /// Encapsulates TileDB Attribute fill value data for storage in Arrow field metadata
 #[derive(Deserialize, Serialize)]
